+++ packages/ifpack2/src/Ifpack2_DenseContainer_decl.hpp
@@ -236,6 +236,9 @@ private:
   //! Copy constructor: Declared but not implemented, to forbid copy construction.
   DenseContainer (const DenseContainer<MatrixType, LocalScalarType>& rhs);
 
+  //! Populate the diagonal blocks
+  void extract();
+
   /// \brief Factor the extracted submatrix.
   ///
   /// Call this after calling extract().
